[KAIZEN-0] maskere enhet og veileders ident i prometheus
